feat: Implement log viewer and report viewer providers with search and notification functionalities
